requirements.txt
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
discord-interactions==0.4.0
discord-py-interactions==5.10.0
discord-typings==0.7.0
discord.py==2.3.2
emoji==2.8.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
panda==0.3.1
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.31.0
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
